requirements.txt
dash
dash-bootstrap-components
pandas
pandas-ta
numpy
yfinance
plotly
gunicorn
psycopg2-binary
python-dotenv